[CDG-60][CDG-61] - Create AdminUserRelation endpoint

+++ src/main/java/com/carolribeiro2112/controle_gastos_backend/controllers/TransactionController.java
@@ -1,14 +1,17 @@
 package com.carolribeiro2112.controle_gastos_backend.controllers;
 
+import com.carolribeiro2112.controle_gastos_backend.domain.adminUserRelation.AdminUserRelation;
 import com.carolribeiro2112.controle_gastos_backend.domain.transaction.Transaction;
 import com.carolribeiro2112.controle_gastos_backend.domain.transaction.TransactionDTO;
 import com.carolribeiro2112.controle_gastos_backend.domain.transaction.TransactionResponseDTO;
+import com.carolribeiro2112.controle_gastos_backend.repositories.AdminUserRelationRepository;
 import com.carolribeiro2112.controle_gastos_backend.services.TransactionService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -19,8 +22,29 @@ public class TransactionController {
     @Autowired
     private TransactionService transactionService;
 
+    @Autowired
+    private AdminUserRelationRepository relationRepository;
+
+    private void verifyAdminAccess(String adminId, String userId) {
+        if(adminId.equals(userId)) {
+            return;
+        }
+        List<AdminUserRelation> relations = relationRepository.findByAdminId(adminId);
+        boolean isAdminOfUser = relations.stream()
+                .anyMatch(relation -> relation.getUser().getId().equals(userId));
+
+        if (!isAdminOfUser) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Admin não autorizado para este usuário");
+        }
+    }
+
     @PostMapping
-    public  ResponseEntity<Transaction> createTransaction(@Valid @RequestBody TransactionDTO transactionDTO) {
+    public  ResponseEntity<Transaction> createTransaction(
+            @RequestHeader(value = "adminId", required = false) String adminId,
+            @Valid @RequestBody TransactionDTO transactionDTO
+    ) {
+        verifyAdminAccess(adminId, transactionDTO.userId());
+
         Transaction transaction = transactionService.saveTransaction(
                 transactionDTO.userId(),
                 transactionDTO.type(),
@@ -31,13 +55,24 @@ public class TransactionController {
     }
 
     @GetMapping
-    public ResponseEntity<List<TransactionResponseDTO>> getAllTransactionsByUser(@RequestParam String userId) {
+    public ResponseEntity<List<TransactionResponseDTO>> getAllTransactionsByUser(
+            @RequestHeader(value = "adminId", required = false) String adminId,
+            @RequestParam String userId
+    ) {
+        verifyAdminAccess(adminId, userId);
+
         List<TransactionResponseDTO> transaction = transactionService.getAllTransactionsByUserId(userId);
         return ResponseEntity.ok(transaction);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTransaction(@PathVariable String id) {
+    public ResponseEntity<Void> deleteTransaction(
+            @RequestHeader(value = "adminId", required = false) String adminId,
+            @PathVariable String id
+    ) {
+        Transaction transaction = transactionService.getTransactionById(id);
+        verifyAdminAccess(adminId, transaction.getUser().getId());
+
         transactionService.deleteTransactionById(id);
         return ResponseEntity.noContent().build();
     }
